Extract common Webpack config from dev and prd
The dev and prd configs shared almost all of their values, so we can
extract those parts of the config to the common config file.
